test/requirements.txt
arrow==0.15.8
astroid==2.4.2
autopep8==1.5.3
click==7.1.2
Faker==4.1.1
flake8==3.8.3
Flask==1.1.2
inflect==4.1.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
python-dateutil==2.8.1
rope==0.17.0
six==1.15.0
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
Werkzeug==1.0.1
wrapt==1.12.1
